Enforce usage of versioned Python interpreter in Makefiles
The Makefiles for dnfdaemon previously assumed that '/usr/bin/python'
pointed to Python 2. However, as support for Python 2 winds down
over the next two years, this is not guaranteed to remain true.

Consequently, Fedora has split out '/usr/bin/python' from the 'python2'
package in order to enforce that packages do not use '/usr/bin/python'
by default. Since the assumption was wrong in the Makefiles to begin
with, it is better to go ahead and fix it in the Makefiles.

Reference: https://fedoraproject.org/wiki/Changes/Move_usr_bin_python_into_separate_package